changed routing name, fixed quiz page and template
