--- pages/auth/TEMPLATE.js
@@ -1,34 +1,0 @@
-import Head from "next/head";
-import Link from "next/link";
-import { Nav } from "../../components/Nav.jsx";
-import {
-  useAuthUser,
-  withAuthUser,
-  withAuthUserTokenSSR,
-  AuthAction,
-} from "next-firebase-auth";
-
-function Template() {
-  const AuthUser = useAuthUser();
-  return (
-    <>
-      <div>
-        <Head>
-          <title>Template</title>
-          <meta name="description" content="Generated by create next app" />
-          <link rel="icon" href="/favicon.ico" />
-        </Head>
-        <Nav email={AuthUser.email} signOut={AuthUser.signOut} />
-
-        <main className="container h-screen mx-auto flex"></main>
-      </div>
-    </>
-  );
-}
-
-export const getServerSideProps = withAuthUserTokenSSR()();
-
-export default withAuthUser({
-  whenUnauthedAfterInit: AuthAction.REDIRECT_TO_LOGIN,
-  authPageURL: "/",
-})(Template);
