--- pages/auth/quiz.js
@@ -1,38 +1,0 @@
-import Head from "next/head";
-import Link from "next/link";
-import { Nav } from "../../components/Nav.jsx";
-import {
-  useAuthUser,
-  withAuthUser,
-  withAuthUserTokenSSR,
-  AuthAction,
-} from "next-firebase-auth";
-
-function Quiz() {
-  const AuthUser = useAuthUser();
-  return (
-    <>
-      <div>
-        <Head>
-          <title>Quizzes</title>
-          <meta name="description" content="Generated by create next app" />
-          <link rel="icon" href="/favicon.ico" />
-        </Head>
-        <Nav email={AuthUser.email} signOut={AuthUser.signOut} />
-
-        <main className="container h-screen mx-auto flex">
-          <h1 className="text-center m-auto text-2xl font-bold">
-            Welcome to Quiz
-          </h1>
-        </main>
-      </div>
-    </>
-  );
-}
-
-export const getServerSideProps = withAuthUserTokenSSR()();
-
-export default withAuthUser({
-  whenUnauthedAfterInit: AuthAction.REDIRECT_TO_LOGIN,
-  authPageURL: "/",
-})(Quiz);
